[ADF-4128] - refractor complete task directive
