+++ demo-shell/src/app/components/app-layout/cloud/task-details-cloud-demo.component.html
@@ -1,8 +1,14 @@
-
 <h4 data-automation-id="task-details-header">Simple page to show the taskId: {{ taskId }} of the app: {{ appName }}</h4>
 
-<adf-cloud-task-header
-    [appName]="appName"
-    [taskId]="taskId"
-    [readOnly]="readOnly">
-</adf-cloud-task-header>
+<div class="adf-task-detail-container">
+    <div class="adf-task-control">
+        <button mat-button (click)="goBack()">Cancel</button>
+        <button mat-button color="primary" *ngIf="isTaskValid() && canCompleteTask()" adf-cloud-complete-task [appName]="appName"
+            [taskId]="taskId"
+            (success)="onCompletedTask($event)">{{ 'ADF_TASK_LIST.DETAILS.BUTTON.COMPLETE' | translate }}</button>
+    </div>
+
+    <adf-cloud-task-header class="adf-demop-card-container" [appName]="appName" [taskId]="taskId" [readOnly]="readOnly">
+    </adf-cloud-task-header>
+
+</div>
