+++ demo-shell/src/app/components/app-layout/cloud/task-details-cloud-demo.component.ts
@@ -15,20 +15,26 @@
  * limitations under the License.
  */
 
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { TaskCloudService, TaskDetailsCloudModel } from '@alfresco/adf-process-services-cloud';
 
 @Component({
     templateUrl: './task-details-cloud-demo.component.html',
     styleUrls: ['./task-details-cloud-demo.component.scss']
 })
-export class TaskDetailsCloudDemoComponent {
+export class TaskDetailsCloudDemoComponent implements OnInit {
 
+    taskDetails: TaskDetailsCloudModel;
     taskId: string;
     appName: string;
     readOnly = false;
 
-    constructor(private route: ActivatedRoute, private router: Router) {
+    constructor(
+        private route: ActivatedRoute,
+        private router: Router,
+        private taskCloudService: TaskCloudService
+        ) {
         this.route.params.subscribe((params) => {
             this.taskId = params.taskId;
         });
@@ -37,8 +43,30 @@ export class TaskDetailsCloudDemoComponent {
         });
     }
 
-    onGoBack() {
-        this.router.navigate([`/cloud/${this.appName}/`]);
+    ngOnInit() {
+        this.loadTaskDetailsById(this.appName, this.taskId);
+    }
 
+    loadTaskDetailsById(appName: string, taskId: string): any {
+        this.taskCloudService.getTaskById(appName, taskId).subscribe(
+            (taskDetails) => {
+                this.taskDetails = taskDetails;
+            });
+    }
+
+    isTaskValid() {
+        return this.appName && this.taskId;
+    }
+
+    canCompleteTask() {
+        return this.taskDetails && this.taskCloudService.canCompleteTask(this.taskDetails);
+    }
+
+    goBack() {
+        this.router.navigate([`/cloud/${this.appName}/`]);
+    }
+
+    onCompletedTask(evt: any) {
+        this.goBack();
     }
 }
