+++ docs/process-services-cloud/complete-task.directive.md
@@ -1,5 +1,5 @@
 ---
-Title: Task Cloud Service
+Title: Complete Cloud Task
 Added: v3.1.0
 Status: Experimental
 Last reviewed: 2019-02-28
@@ -12,7 +12,7 @@ Complete a task
 ## Basic Usage
 
 ```html
-<button adf-complete-task [appName]="appName" [taskId]="taskId" (taskCompleted)="onTaskCompleted()">Complete</button>
+<button adf-cloud-complete-task [appName]="appName" [taskId]="taskId" (success)="onTaskCompleted()">Complete</button>
 ```
 ## Class members
 
@@ -22,5 +22,5 @@ Complete a task
 | ---- | ---- | ------------- | ----------- |
 | taskId | `string` | empty |(Required) The id of the task. |
 | appName | `string` | empty | (Required) The name of the application. |
-| taskCompleted | `EventEmitter<any>` | empty | Emitted when the task is completed. |
+| success | `EventEmitter<any>` | empty | Emitted when the task is completed. |
 | error | `EventEmitter<any>` | empty | Emitted when the task cannot be completed.  |
