+++ lib/process-services-cloud/src/lib/task/task-header/components/task-header-cloud.component.html
@@ -1,19 +1,11 @@
 <h3 class="adf-task-tiitle">{{ taskDetails.name }}</h3>
 
 <div class="adf-task-header-container">
-    <div class="adf-task-control">
-        <button mat-button (click)="goBack()">Cancel</button>
-        <button mat-button color="primary" *ngIf="isTaskValid() && canCompleteTask()" adf-complete-task
-            [appName]="appName" [taskId]="taskId"
-            (success)="onCompletedTask($event)"
-            >{{ 'ADF_TASK_LIST.DETAILS.BUTTON.COMPLETE' | translate }}</button>
-    </div>
-
     <mat-card *ngIf="isTaskValid()" class="adf-card-container">
         <mat-card-content>
             <adf-card-view
                 [properties]="properties"
-                [editable]="!isCompleted()">
+                [editable]="!taskDetails.isCompleted()">
             </adf-card-view>
         </mat-card-content>
