+++ lib/process-services-cloud/src/lib/task/task-header/components/task-header-cloud.component.scss
@@ -32,23 +32,3 @@
         }
     }
 }
-
-.adf {
-
-    &-task-header-container {
-        display: flex;
-    }
-
-    &-task-tiitle {
-        margin-left:15px;
-    }
-
-    &-task-control {
-        width:70%;
-    }
-
-    &-card-container {
-        width:30%;
-        font-family: inherit;
-    }
-}
