+++ lib/process-services-cloud/src/lib/task/task-header/services/task-cloud.service.ts
@@ -16,22 +16,29 @@
  */
 
 import { Injectable } from '@angular/core';
-import { AlfrescoApiService, LogService, AppConfigService, StorageService } from '@alfresco/adf-core';
-import { from, throwError } from 'rxjs';
-import { catchError } from 'rxjs/operators';
-import { TaskDetailsCloudModel } from '../../start-task/models/task-details-cloud.model';
+import { AlfrescoApiService, LogService, AppConfigService, IdentityUserService } from '@alfresco/adf-core';
+import { from, throwError, Observable } from 'rxjs';
+import { catchError, map } from 'rxjs/operators';
+import { TaskDetailsCloudModel, TaskStatusEnum } from '../../start-task/models/task-details-cloud.model';
 
 @Injectable({
     providedIn: 'root'
 })
 export class TaskCloudService {
 
+    contextRoot: string;
+    contentTypes = ['application/json'];
+    accepts = ['application/json'];
+    returnType = Object;
+
     constructor(
         private apiService: AlfrescoApiService,
         private appConfigService: AppConfigService,
         private logService: LogService,
-        private storage: StorageService
-    ) { }
+        private identityUserService: IdentityUserService
+    ) {
+        this.contextRoot = this.appConfigService.get('bpmHost', '');
+    }
 
     /**
      * Complete a task.
@@ -63,16 +70,132 @@ export class TaskCloudService {
      * @returns Boolean value if the task can be completed
      */
     canCompleteTask(taskDetails: TaskDetailsCloudModel): boolean {
-        const currentUser = this.storage.getItem('USERNAME');
+        const currentUser = this.identityUserService.getCurrentUserInfo().username;
         return taskDetails.owner === currentUser && !this.isCompleted(taskDetails);
     }
 
     private isCompleted(taskDetails: TaskDetailsCloudModel) {
-        return taskDetails && taskDetails.status && taskDetails.status.toLowerCase() === 'completed';
+        return taskDetails && taskDetails.status && taskDetails.status.toUpperCase() === TaskStatusEnum.COMPLETED;
     }
 
     private buildCompleteTaskUrl(appName: string, taskId: string): any {
         return `${this.appConfigService.get('bpmHost')}/${appName}-rb/v1/tasks/${taskId}/complete`;
+    }
+
+    /**
+     * Gets details of a task.
+     * @param appName Name of the app
+     * @param taskId ID of the task whose details you want
+     * @returns Task details
+     */
+    getTaskById(appName: string, taskId: string): Observable<TaskDetailsCloudModel> {
+        if (appName && taskId) {
+
+            let queryUrl = `${this.contextRoot}/${appName}-query/v1/tasks/${taskId}`;
+            return from(this.apiService.getInstance()
+                .oauth2Auth.callCustomApi(queryUrl, 'GET',
+                    null, null, null,
+                    null, null,
+                    this.contentTypes, this.accepts,
+                    this.returnType, null, null)
+            ).pipe(
+                map((res: any) => {
+                    return new TaskDetailsCloudModel(res.entry);
+                }),
+                catchError((err) => this.handleError(err))
+            );
+        } else {
+            this.logService.error('AppName and TaskId are mandatory for querying a task');
+            return throwError('AppName/TaskId not configured');
+        }
+    }
+
+    /**
+     * Updates the details (name, description, due date) for a task.
+     * @param appName Name of the app
+     * @param taskId ID of the task to update
+     * @param updatePayload Data to update the task
+     * @returns Updated task details
+     */
+    updateTask(appName: string, taskId: string, updatePayload: any): any {
+        if (appName && taskId) {
+
+            updatePayload.payloadType = 'UpdateTaskPayload';
+
+            let queryUrl = `${this.contextRoot}/${appName}-rb/v1/tasks/${taskId}`;
+            return from(this.apiService.getInstance()
+                .oauth2Auth.callCustomApi(queryUrl, 'PUT',
+                    null, null, null,
+                    null, updatePayload,
+                    this.contentTypes, this.accepts,
+                    this.returnType, null, null)
+            ).pipe(
+                map((res: any) => {
+                    return new TaskDetailsCloudModel(res.entry);
+                }),
+                catchError((err) => this.handleError(err))
+            );
+        } else {
+            this.logService.error('AppName and TaskId are mandatory for querying a task');
+            return throwError('AppName/TaskId not configured');
+        }
+    }
+
+    /**
+     * Claims a task for an assignee.
+     * @param appName Name of the app
+     * @param taskId ID of the task to claim
+     * @param assignee User to assign the task to
+     * @returns Details of the claimed task
+     */
+    claimTask(appName: string, taskId: string, assignee: string): any {
+        if (appName && taskId) {
+
+            let queryUrl = `${this.contextRoot}/${appName}-rb/v1/tasks/${taskId}/claim?assignee=${assignee}`;
+            return from(this.apiService.getInstance()
+                .oauth2Auth.callCustomApi(queryUrl, 'POST',
+                    null, null, null,
+                    null, null,
+                    this.contentTypes, this.accepts,
+                    this.returnType, null, null)
+            ).pipe(
+                map((res: any) => {
+                    return new TaskDetailsCloudModel(res.entry);
+                }),
+                catchError((err) => this.handleError(err))
+            );
+        } else {
+            this.logService.error('AppName and TaskId are mandatory for querying a task');
+            return throwError('AppName/TaskId not configured');
+        }
+    }
+
+    /**
+     * Un-claims a task.
+     * @param appName Name of the app
+     * @param taskId ID of the task to unclaim
+     * @returns Details of the task that was unclaimed
+     */
+    unclaimTask(appName: string, taskId: string): any {
+        if (appName && taskId) {
+
+            let queryUrl = `${this.contextRoot}/${appName}-rb/v1/tasks/${taskId}/release`;
+            return from(this.apiService.getInstance()
+                .oauth2Auth.callCustomApi(queryUrl, 'POST',
+                    null, null, null,
+                    null, null,
+                    this.contentTypes, this.accepts,
+                    this.returnType, null, null)
+            ).pipe(
+                map((res: any) => {
+                    return new TaskDetailsCloudModel(res.entry);
+                }),
+                catchError((err) => this.handleError(err))
+            );
+        } else {
+            this.logService.error('AppName and TaskId are mandatory for querying a task');
+            return throwError('AppName/TaskId not configured');
+        }
     }
 
     private handleError(error: any) {
