--- lib/process-services-cloud/src/lib/task/task-header/services/task-header-cloud.service.spec.ts
@@ -1,245 +1,0 @@
-/*!
- * @license
- * Copyright 2019 Alfresco Software, Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { async } from '@angular/core/testing';
-import { setupTestBed } from '@alfresco/adf-core';
-import { AlfrescoApiServiceMock, LogService, AppConfigService, StorageService, CoreModule } from '@alfresco/adf-core';
-import { fakeTaskDetailsCloud } from '../mocks/fake-task-details-response.mock';
-import { TaskHeaderCloudService } from './task-header-cloud.service';
-
-describe('Task Header Cloud Service', () => {
-
-    let service: TaskHeaderCloudService;
-    let alfrescoApiMock: AlfrescoApiServiceMock;
-
-    function returnFakeTaskDetailsResults() {
-        return {
-            oauth2Auth: {
-                callCustomApi : () => {
-                    return Promise.resolve(fakeTaskDetailsCloud);
-                }
-            }
-        };
-    }
-
-    setupTestBed({
-        imports: [
-            CoreModule.forRoot()
-        ]
-    });
-
-    beforeEach(async(() => {
-        alfrescoApiMock = new AlfrescoApiServiceMock(new AppConfigService(null), new StorageService() );
-        service = new TaskHeaderCloudService(alfrescoApiMock,
-                                           new AppConfigService(null),
-                                           new LogService(new AppConfigService(null)));
-    }));
-
-    it('should return the task details when querying by id', (done) => {
-        const appName = 'taskp-app';
-        const taskId = '68d54a8f';
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.getTaskById(appName, taskId).subscribe((res: any) => {
-            expect(res).toBeDefined();
-            expect(res).not.toBeNull();
-            expect(res.appName).toBe('task-app');
-            expect(res.id).toBe('68d54a8f');
-            done();
-        });
-    });
-
-    it('should throw error if appName is not defined when querying by id', (done) => {
-        const appName = null;
-        const taskId = '68d54a8f';
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.getTaskById(appName, taskId).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should throw error if taskId is not defined when querying by id', (done) => {
-        const appName = 'task-app';
-        const taskId = null;
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.getTaskById(appName, taskId).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should return the task details when updating a task', (done) => {
-        const appName = 'taskp-app';
-        const taskId = '68d54a8f';
-        const updatePayload = { description: 'New description' };
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.updateTask(appName, taskId, updatePayload).subscribe((res: any) => {
-            expect(res).toBeDefined();
-            expect(res).not.toBeNull();
-            expect(res.appName).toBe('task-app');
-            expect(res.id).toBe('68d54a8f');
-            done();
-        });
-    });
-
-    it('should throw error if appName is not defined when updating a task', (done) => {
-        const appName = null;
-        const taskId = '68d54a8f';
-        const updatePayload = { description: 'New description' };
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.updateTask(appName, taskId, updatePayload).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should throw error if taskId is not defined when updating a task', (done) => {
-        const appName = 'task-app';
-        const taskId = null;
-        const updatePayload = { description: 'New description' };
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.updateTask(appName, taskId, updatePayload).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should return the task details when updating a task', (done) => {
-        const appName = 'taskp-app';
-        const taskId = '68d54a8f';
-        const updatePayload = { description: 'New description' };
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.updateTask(appName, taskId, updatePayload).subscribe((res: any) => {
-            expect(res).toBeDefined();
-            expect(res).not.toBeNull();
-            expect(res.appName).toBe('task-app');
-            expect(res.id).toBe('68d54a8f');
-            done();
-        });
-    });
-
-    it('should throw error if appName is not defined when querying by id', (done) => {
-        const appName = null;
-        const taskId = '68d54a8f';
-        const updatePayload = { description: 'New description' };
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.updateTask(appName, taskId, updatePayload).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should throw error if taskId is not defined updating a task', (done) => {
-        const appName = 'task-app';
-        const taskId = null;
-        const updatePayload = { description: 'New description' };
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.updateTask(appName, taskId, updatePayload).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should return the task details when claiming a task', (done) => {
-        const appName = 'taskp-app';
-        const assignee = 'user12';
-        const taskId = '68d54a8f';
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.claimTask(appName, taskId, assignee).subscribe((res: any) => {
-            expect(res).toBeDefined();
-            expect(res).not.toBeNull();
-            expect(res.appName).toBe('task-app');
-            expect(res.id).toBe('68d54a8f');
-            done();
-        });
-    });
-
-    it('should throw error if appName is not defined when claiming a task', (done) => {
-        const appName = null;
-        const taskId = '68d54a8f';
-        const assignee = 'user12';
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.claimTask(appName, taskId, assignee).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should throw error if taskId is not defined when claiming a task', (done) => {
-        const appName = 'task-app';
-        const taskId = null;
-        const assignee = 'user12';
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.claimTask(appName, taskId, assignee).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should return the task details when unclaiming a task', (done) => {
-        const appName = 'taskp-app';
-        const taskId = '68d54a8f';
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.unclaimTask(appName, taskId).subscribe((res: any) => {
-            expect(res).toBeDefined();
-            expect(res).not.toBeNull();
-            expect(res.appName).toBe('task-app');
-            expect(res.id).toBe('68d54a8f');
-            done();
-        });
-    });
-
-    it('should throw error if appName is not defined when unclaiming a task', (done) => {
-        const appName = null;
-        const taskId = '68d54a8f';
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.unclaimTask(appName, taskId).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-
-    it('should throw error if taskId is not defined when unclaiming a task', (done) => {
-        const appName = 'task-app';
-        const taskId = null;
-        spyOn(alfrescoApiMock, 'getInstance').and.callFake(returnFakeTaskDetailsResults);
-        service.unclaimTask(appName, taskId).subscribe(
-            () => { },
-            (error) => {
-                expect(error).toBe('AppName/TaskId not configured');
-                done();
-            });
-    });
-});
