+++ lib/process-services-cloud/src/lib/task/task-header/task-header-cloud.module.ts
@@ -20,8 +20,6 @@ import { CommonModule } from '@angular/common';
 import { MaterialModule } from '../../material.module';
 import { DataTableModule, TemplateModule, CardViewModule, CoreModule } from '@alfresco/adf-core';
 import { TaskHeaderCloudComponent } from './components/task-header-cloud.component';
-import { TaskHeaderCloudService } from './services/task-header-cloud.service';
-import { TaskDirectiveModule } from '../directives/task-directive.module';
 
 @NgModule({
     imports: [
@@ -30,17 +28,13 @@ import { TaskDirectiveModule } from '../directives/task-directive.module';
         DataTableModule,
         TemplateModule,
         CardViewModule,
-        CoreModule,
-        TaskDirectiveModule
+        CoreModule
     ],
     declarations: [
         TaskHeaderCloudComponent
     ],
     exports: [
         TaskHeaderCloudComponent
-    ],
-    providers: [
-        TaskHeaderCloudService
     ]
 })
 export class TaskHeaderCloudModule { }
